keep scan options in scan.py
